+++ src/main/java/com/github/jlangch/venice/impl/functions/IOFnBlacklisted.java
@@ -41,6 +41,7 @@ public class IOFnBlacklisted {
 					"load-file",
 					"load-classpath-file",
 					"*load-file",
+					"*load-resource",
 					"*load-classpath-file",
 					
 					// classloader